Workaround multicore issue with Fiber unittest
Fiber context switching fails badly when trading a Fiber between threads
on a multicore system.  Does not happen when LDC optimizer is off -O0,
so suspect that a read or write is being optimized away.  To allow most
of the unittest to pass, define version WIP_FiberIssue when compiling
with optimization levels -O1 or better.